Move nadlink.py to __init__.py
